Work in progress

+++ app/models/concerns/current_cart.rb
@@ -1,14 +1,14 @@
 module CurrentCart
 
     def set_cart
-        if !current_user.nil?
+        if current_user.present?
             if session[:cart_id].present?
                 @cart = Cart.find(session[:cart_id])
-                if !@cart.order_items.first.nil?
-                    if @cart.user_id.nil?
+                if !@cart.order_items.first.nil? && @cart.user_id.nil?
+                    if Cart.where(user_id: current_user.id).last.present?
                         Cart.where(user_id: current_user.id).last.destroy
-                        @cart.update_attributes(user_id: current_user.id)
                     end
+                    @cart.update_attributes(user_id: current_user.id)
                 else
                     if @cart.user_id.nil?
                         @cart.destroy
@@ -21,19 +21,20 @@ module CurrentCart
                 @cart = Cart.create(user_id: current_user.id)
                 session[:cart_id] = @cart.id
             end
-
+            
         else
             if session[:cart_id].present?
-                if !Cart.all.last.nil?
+                begin
                     @cart = Cart.find(session[:cart_id])
-                else
-                    @cart=Cart.create
-                    session[:cart_id]=@cart.id
+                rescue
+                    @cart = Cart.create
+                    session[:cart_id] = @cart.id
                 end
             else
                 @cart = Cart.create
                 session[:cart_id] = @cart.id
             end
+            
         end
     rescue ActiveRecord::RecordNotFound
         @cart = Cart.create(user_id: current_user.id)
